refactor: move functions the usePoolContract for poolUtils file
