+++ src/hooks/usePoolContract.tsx
@@ -159,67 +159,6 @@ function PoolContract(contract: Contract) {
     return new BigNumber(value)
   }
 
-  const checkTokenWithHigherLiquidityPool = async (underlyingAssets: underlyingAssetsInfo[]) => {
-
-    const tokensNormalized = await Promise.all(
-      underlyingAssets.map(async (item) => {
-        const normalizedWeightToken = await normalizedWeight(item.token.id);
-
-        return {
-          address: item.token.id,
-          normalizedWeight: normalizedWeightToken
-        }
-      })
-    )
-
-    const tokensNormalizedWithSort = (tokensNormalized).sort(function (a, b) {
-      if (a.normalizedWeight > b.normalizedWeight) {
-        return -1;
-      }
-      if (a.normalizedWeight < b.normalizedWeight) {
-        return 1;
-      }
-      return 0;
-    });
-
-    return tokensNormalizedWithSort[0] || {}
-  }
-
-  const getTokenWrapped = (underlyingAssets: underlyingAssetsInfo[], address: string) => {
-    const tokenAddresses = underlyingAssets.find(item =>
-      item.token.id.toLocaleLowerCase() === address.toLocaleLowerCase()
-    )
-
-    if (tokenAddresses?.token.wraps) {
-      return tokenAddresses?.token.wraps.id
-    }
-
-    return address
-  }
-
-  const checkTokenInThePool = async (underlyingAssets: underlyingAssetsInfo[], address: string) => {
-    const tokensChecked = underlyingAssets.map(item => {
-      if (item.token.is_wrap_token === 1) {
-        return {
-          address: item.token.wraps.id.toLowerCase(),
-          is_wraps: 1,
-          yrt: item.token.id
-        }
-      }
-      return {
-        address: item.token.id.toLowerCase(),
-        is_wraps: 0,
-        yrt: ""
-      }
-    })
-
-    return tokensChecked.find(token => {
-      if (token.address === address.toLowerCase()) {
-        return token
-      }
-    })
-  }
-
   return {
     events,
 
@@ -234,10 +173,7 @@ function PoolContract(contract: Contract) {
     normalizedWeight,
     swapFee,
     exitFee,
-    totalDenormalizedWeight,
-    checkTokenWithHigherLiquidityPool,
-    checkTokenInThePool,
-    getTokenWrapped
+    totalDenormalizedWeight
   }
 }
